// src/components/CalendarBase/hooks/types.ts
export type CalendarLayout = 'days' | 'months' | 'quarters' | 'years';